feat: 改进了对satori协议转义字符的支持
